Split out makeImageURLSafe
