+++ client/lib/post-normalizer/utils/index.js
@@ -7,26 +7,12 @@ import url from 'url';
 /**
  * Internal Dependencies
  */
-import safeImageURL from 'lib/safe-image-url';
 import { isPhotonHost } from 'lib/post-normalizer/utils/is-photon-host';
-import { maxWidthPhotonishURL } from 'lib/post-normalizer/utils/max-width-photonish-url';
 
 export { isPhotonHost } from 'lib/post-normalizer/utils/is-photon-host';
 export { imageSizeFromAttachments } from 'lib/post-normalizer/utils/image-size-from-attachments';
 export { maxWidthPhotonishURL } from 'lib/post-normalizer/utils/max-width-photonish-url';
-
-export function makeImageURLSafe( object, propName, maxWidth, baseURL ) {
-	if ( object && object[ propName ] ) {
-		if ( baseURL && ! url.parse( object[ propName ], true, true ).hostname ) {
-			object[ propName ] = url.resolve( baseURL, object[ propName ] );
-		}
-		object[ propName ] = safeImageURL( object[ propName ] );
-
-		if ( maxWidth ) {
-			object[ propName ] = maxWidthPhotonishURL( object[ propName ], maxWidth );
-		}
-	}
-}
+export { makeImageURLSafe } from 'lib/post-normalizer/utils/make-image-url-safe';
 
 export function domForHtml( html ) {
 	if ( typeof DOMParser !== 'undefined' && window.DOMParser.prototype.parseFromString ) {
